+++ magda-gateway/src/createAuthRouter.ts
@@ -14,10 +14,14 @@ export interface AuthRouterOptions {
     aafClientSecret: string;
     arcgisClientId: string;
     arcgisClientSecret: string;
+    arcgisInstanceBaseUrl: string;
     ckanUrl: string;
     authorizationApi: string;
     externalUrl: string;
     userId: string;
+    vanguardWsFedIdpUrl: string;
+    vanguardWsFedRealm: string;
+    vanguardWsFedCertificate: string;
 }
 
 export default function createAuthRouter(options: AuthRouterOptions): Router {
@@ -65,6 +69,7 @@ export default function createAuthRouter(options: AuthRouterOptions): Router {
                 passport: passport,
                 clientId: options.arcgisClientId,
                 clientSecret: options.arcgisClientSecret,
+                arcgisInstanceBaseUrl: options.arcgisInstanceBaseUrl,
                 externalAuthHome: `${options.externalUrl}/auth`
             })
         },
@@ -87,6 +92,18 @@ export default function createAuthRouter(options: AuthRouterOptions): Router {
                 aafClientUri: options.aafClientUri,
                 aafClientSecret: options.aafClientSecret,
                 externalUrl: options.externalUrl
+            })
+        },
+        {
+            id: "vanguard",
+            enabled: options.vanguardWsFedIdpUrl ? true : false,
+            authRouter: require("./oauth2/vanguard").default({
+                authorizationApi: authApi,
+                passport: passport,
+                wsFedIdpUrl: options.vanguardWsFedIdpUrl,
+                wsFedRealm: options.vanguardWsFedRealm,
+                wsFedCertificate: options.vanguardWsFedCertificate,
+                externalAuthHome: `${options.externalUrl}/auth`
             })
         }
     ];