+++ magda-typescript-common/src/OpaCompileResponseParser.ts
@@ -318,6 +318,63 @@ export class RegoTerm {
     }
 
     /**
+     * Return RegoRef instance if this term is a RegoRef.
+     *
+     * @returns {RegoRef}
+     * @memberof RegoTerm
+     */
+    getRef(): RegoRef {
+        if (this.isRef()) {
+            return this.value as RegoRef;
+        }
+        throw new Error(`Term ${this.asString()} is not a ref`);
+    }
+
+    /**
+     * If the term is a reference and it contains any collection lookup
+     * e.g.
+     * - objectA.propB.collectionC[_]
+     * - objectA.propB.collectionC[_].ABC[_].name
+     * - objectA.propB.collectionC[_].id
+     *
+     * @returns {boolean}
+     * @memberof RegoTerm
+     */
+    hasCollectionLookup(): boolean {
+        if (!this.isRef()) return false;
+        const ref = this.getRef();
+        return ref.hasCollectionLookup();
+    }
+
+    /**
+     * The term is not only a Reference but a reference contains simple collection lookup
+     * i.e. only contains one collection lookup and the whole ref ends with the only collection lookup
+     * e.g. objectA.propB.collectionC[_]
+     * Note: objectA.propB.collectionC[_].name is not a simple collection lookup as it resolve to single value (`name` property)
+     * rather than a collection
+     *
+     * @returns {boolean}
+     * @memberof RegoTerm
+     */
+    isSimpleCollectionLookup(): boolean {
+        if (!this.isRef()) return false;
+        const ref = this.getRef();
+        return ref.isSimpleCollectionLookup();
+    }
+
+    /**
+     *
+     *
+     * @returns {boolean}
+     * @memberof RegoTerm
+     */
+    isResolveAsCollectionValue(): boolean {
+        if (!this.isRef()) return false;
+        const ref = this.getRef();
+        return ref.isResolveAsCollectionValue();
+    }
+
+    /**
      * If it's a reference term, return its full string representation
      * Otherwise, throw exception
      *
@@ -744,7 +801,7 @@ export class RegoRef {
             .forEach(prefix => {
                 if (!prefix) return;
                 const idx = result.indexOf(prefix);
-                if (idx === -1) return;
+                if (idx !== 0) return;
                 result = result.substring(prefix.length);
             });
         return result;
@@ -760,7 +817,7 @@ export class RegoRef {
                 } else {
                     if (part.type == "var") {
                         // --- it's a collection lookup
-                        // --- var name doesn't matter
+                        // --- var name doesn't matter for most cases
                         partStr = "[_]";
                     } else {
                         partStr = part.value;
@@ -771,12 +828,18 @@ export class RegoRef {
                 //--- a.[_].[_] should be a[_][_]
             })
             .join(".")
-            .replace(".[", "[");
+            .replace(/\.\[/g, "[");
         return this.removeAllPrefixs(str, removalPrefixs);
     }
 
     refString(removalPrefixs: string[] = []): string {
         return this.fullRefString(removalPrefixs).replace("\\[_\\]$", "");
+    }
+
+    asCollectionRefs(removalPrefixs: string[] = []): string[] {
+        return this.fullRefString(removalPrefixs)
+            .split("[_]")
+            .map(refStr => refStr.replace(/^\./, ""));
     }
 
     isOperator(): boolean {
@@ -786,7 +849,12 @@ export class RegoRef {
     // --- the first var type won't count as collection lookup
     hasCollectionLookup(): boolean {
         if (this.parts.length <= 1) return false;
-        else return this.parts.findIndex(part => part.type === "var") >= 1;
+        else {
+            return (
+                this.parts.slice(1).findIndex(part => part.type === "var") !==
+                -1
+            );
+        }
     }
 
     // -- simple collection only contains 1 level lookup
@@ -795,10 +863,16 @@ export class RegoRef {
         if (this.parts.length <= 1) return false;
         else {
             return (
-                this.parts.findIndex(part => part.type === "var") ===
-                this.parts.length - 1
+                this.parts.slice(1).findIndex(part => part.type === "var") ===
+                this.parts.length - 2
             );
         }
+    }
+
+    // --- see comment for same name method of RegoTerm
+    isResolveAsCollectionValue(): boolean {
+        const refString = this.fullRefString();
+        return refString.lastIndexOf("[_]") === refString.length - 3;
     }
 
     asOperator(): string {