+++ magda-web-client/src/Components/Common/AddDatasetProgressMeter.tsx
@@ -96,6 +96,7 @@ const AddDatasetProgressMeter = (props: PropsType) => {
         }
         return (
             <div
+                key={idx}
                 className={`col-sm-2 step-item-container ${statusClass}`}
                 onClick={() => {
                     if (statusClass !== "past-item") {
@@ -117,13 +118,15 @@ const AddDatasetProgressMeter = (props: PropsType) => {
 
     return (
         <div className="row add-dataset-progress-meter">
-            <div className="col-sm-2 step-item-heading">
-                <div className="heading">Add a dataset</div>
-            </div>
-            <div className="col-sm-10 step-item-body">
-                {Steps.map((item, idx) =>
-                    renderStepItem(item, idx, currentStep, datasetId)
-                )}
+            <div className="container">
+                <div className="col-sm-2 step-item-heading">
+                    <div className="heading">Add a dataset</div>
+                </div>
+                <div className="col-sm-10 step-item-body">
+                    {Steps.map((item, idx) =>
+                        renderStepItem(item, idx, currentStep, datasetId)
+                    )}
+                </div>
             </div>
         </div>
     );